ctdb: Fix a 32-bit problem

Autobuild-User(master): Michael Adam <obnox@samba.org>
Autobuild-Date(master): Thu Sep  3 22:12:02 CEST 2015 on sn-devel-104

+++ ctdb/server/ctdb_daemon.c
@@ -1083,7 +1083,7 @@ static void ctdb_tevent_trace(enum tevent_trace_point tp,
 			if (diff.tv_sec > 3) {
 				DEBUG(DEBUG_ERR,
 				      ("Handling event took %ld seconds!\n",
-				       diff.tv_sec));
+				       (long)diff.tv_sec));
 			}
 		}
 		tevent_before_wait_ts = now;
@@ -1095,7 +1095,7 @@ static void ctdb_tevent_trace(enum tevent_trace_point tp,
 			if (diff.tv_sec > 3) {
 				DEBUG(DEBUG_CRIT,
 				      ("No event for %ld seconds!\n",
-				       diff.tv_sec));
+				       (long)diff.tv_sec));
 			}
 		}
 		tevent_after_wait_ts = now;
